requirements.txt
python-dotenv
supabase
-e git+https://github.com/google/generative-ai-python.git@main#egg=google-generativeai
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic>=2.0,<3.0